# source/aws-web/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agricultural Weighing System",
        "description": "aws-server",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/weigh/scale": {
            "put": {
                "tags": [
                    "称重模块"
                ],
                "summary": "更新电子秤信息",
                "operationId": "updateScale",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScaleUpdateVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScaleVO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "添加电子秤",
                "operationId": "addScale",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScaleAddVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScaleVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/weigh/record": {
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "添加称重记录",
                "operationId": "addRecord",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordAddVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecordVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/record/todo": {
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "提交待处理称重记录",
                "operationId": "handleTodoRecord",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/record/todo/list": {
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "获取待处理称重记录",
                "operationId": "getTodos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodosGetVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TodoListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/record/list": {
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "获取称重记录",
                "operationId": "getRecords",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordsGetVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecordListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/monitor": {
            "post": {
                "tags": [
                    "称重模块"
                ],
                "summary": "添加称重结果监控者",
                "operationId": "addMonitor",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/weigh/summary": {
            "get": {
                "tags": [
                    "称重模块"
                ],
                "summary": "获取员工各作业采摘量",
                "operationId": "getUserWorkSummary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserWorkOutputVO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/scale/{id}": {
            "get": {
                "tags": [
                    "称重模块"
                ],
                "summary": "获取电子秤",
                "operationId": "getScale",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScaleVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weigh/scale/list": {
            "get": {
                "tags": [
                    "称重模块"
                ],
                "summary": "获取电子秤列表",
                "operationId": "getScales",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScaleListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ScaleUpdateVO": {
                "type": "object",
                "description": "更新电子秤信息表单",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "电子秤编号"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "状态，0 为禁用，1 为启用，2 为已删除"
                    }
                }
            },
            "ScaleVO": {
                "type": "object",
                "description": "电子秤信息",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "电子秤编号"
                    },
                    "sid": {
                        "type": "string",
                        "description": "MQTT CLIENT"
                    },
                    "model": {
                        "type": "string",
                        "description": "型号"
                    },
                    "maxCapacity": {
                        "type": "number",
                        "description": "最大量程"
                    },
                    "minCapacity": {
                        "type": "number",
                        "description": "最小量程"
                    },
                    "unit": {
                        "type": "integer",
                        "format": "int32",
                        "description": "量程单位，0 为 mg，1 为 g，2 为 kg，3 为 t，4 为 lb（磅），5 为 oz（盎司），6 为 ct（克拉）"
                    },
                    "verificationInterval": {
                        "type": "integer",
                        "format": "int32",
                        "description": "检定分度值"
                    },
                    "displayInterval": {
                        "type": "integer",
                        "format": "int32",
                        "description": "显示分度值"
                    },
                    "unitDv": {
                        "type": "integer",
                        "format": "int32",
                        "description": "分度值单位，0 为 mg，1 为 g，2 为 kg，3 为 t，4 为 lb（磅），5 为 oz（盎司），6 为 ct（克拉）"
                    },
                    "createTime": {
                        "type": "integer",
                        "format": "int64",
                        "description": "创建时间，毫秒级时间戳"
                    },
                    "updateTime": {
                        "type": "integer",
                        "format": "int64",
                        "description": "更新时间，毫秒级时间戳"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "状态，0 为禁用，1 为启用，2 为已删除"
                    }
                }
            },
...
            "ScaleAddVO": {
                "type": "object",
                "description": "新增电子秤信息表单",
                "properties": {
                    "skey": {
                        "type": "string",
                        "description": "电子秤密钥"
                    },
                    "model": {
                        "type": "string",
                        "description": "型号"
                    },
                    "maxCapacity": {
                        "type": "number",
                        "description": "最大量程"
                    },
                    "minCapacity": {
                        "type": "number",
                        "description": "最小量程"
                    },
                    "unit": {
                        "type": "integer",
                        "format": "int32",
                        "description": "量程单位，0 为 mg，1 为 g，2 为 kg，3 为 t，4 为 lb（磅），5 为 oz（盎司），6 为 ct（克拉）"
                    },
                    "verificationInterval": {
                        "type": "integer",
                        "format": "int32",
                        "description": "检定分度值"
                    },
                    "displayInterval": {
                        "type": "integer",
                        "format": "int32",
                        "description": "显示分度值"
                    },
                    "unitDv": {
                        "type": "integer",
                        "format": "int32",
                        "description": "分度值单位，0 为 mg，1 为 g，2 为 kg，3 为 t，4 为 lb（磅），5 为 oz（盎司），6 为 ct（克拉）"
                    }
                }
            },
            "RecordAddVO": {
                "type": "object",
                "description": "新增称重记录表单",
                "properties": {
                    "produceId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "果实 Id"
                    },
                    "produceName": {
                        "type": "string",
                        "description": "果实名称"
                    },
                    "image": {
                        "type": "string",
                        "description": "采摘作业图片 URL"
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "员工编号"
                    },
                    "scaleId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "电子秤编号"
                    },
                    "dataValue": {
                        "type": "number",
                        "description": "称重结果"
                    },
                    "dataErrorMargin": {
                        "type": "number",
                        "description": "称量结果误差范围，正负多少"
                    },
                    "unit": {
                        "type": "integer",
                        "format": "int32",
                        "description": "称重单位，0 为 mg，1 为 g，2 为 kg，3 为 t，4 为 lb（磅），5 为 oz（盎司），6 为 ct（克拉"
                    },
                    "dataTime": {
                        "type": "integer",
                        "format": "int64",
                        "description": "称重时间，毫秒级时间戳"
                    }
                }
            },
...
            "TodosGetVO": {
                "type": "object",
                "description": "待处理称重记录查询参数",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "description": "当前页码"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "每页记录数"
                    }
                }
            },
...
        }
    }
}